refactor: テストファイルを規約に従って分割
- normal-cases.test.ts: 配列化テストケースのみに
- immutability.test.ts: イミュータブル性テスト
- property-preservation.test.ts: プロパティ維持テスト

規約「1ファイルに配列化テストケースは1つまで」に準拠
