+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/RewriteRule/withActive/normal-cases.test.ts
@@ -1,16 +1,15 @@
 /**
- * RewriteRule.withActive - 正常系テスト
+ * RewriteRule.withActive - 正常系テスト（isActive状態変更）
  * 1. isActive=true のルールを false に変更できる
  * 2. isActive=false のルールを true に変更できる
- * 3. 同じ値を設定しても新しいインスタンスが返される
- * 4. 元のインスタンスは変更されない（イミュータブル性）
- * 5. 他のプロパティは維持される
+ * 3. 同じ値(true)を設定しても新しいインスタンスが返される
+ * 4. 同じ値(false)を設定しても新しいインスタンスが返される
  */
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 
-describe('RewriteRule.withActive - 正常系', () => {
+describe('RewriteRule.withActive - 正常系（isActive状態変更）', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
@@ -68,53 +67,5 @@ describe('RewriteRule.withActive - 正常系', () => {
       expect(newRule.isActive).toBe(testCase.input.newIsActive);
       expect(newRule).not.toBe(originalRule);
     });
-  });
-
-  it('元のインスタンスは変更されない（イミュータブル性）', () => {
-    const ruleId = 1;
-    const oldString = 'old';
-    const newString = 'new';
-    const urlPattern = 'https://example.com';
-    const isRegex = false;
-    const initialIsActive = true;
-
-    const originalRule = new RewriteRule(
-      ruleId,
-      oldString,
-      newString,
-      urlPattern,
-      isRegex,
-      initialIsActive
-    );
-
-    originalRule.withActive(false);
-
-    expect(originalRule.isActive).toBe(true);
-  });
-
-  it('他のプロパティは維持される', () => {
-    const ruleId = 42;
-    const oldString = 'oldValue';
-    const newString = 'newValue';
-    const urlPattern = 'https://test.example.com/path';
-    const isRegex = true;
-    const initialIsActive = true;
-
-    const originalRule = new RewriteRule(
-      ruleId,
-      oldString,
-      newString,
-      urlPattern,
-      isRegex,
-      initialIsActive
-    );
-
-    const newRule = originalRule.withActive(false);
-
-    expect(newRule.id).toBe(ruleId);
-    expect(newRule.oldString).toBe(oldString);
-    expect(newRule.newString).toBe(newString);
-    expect(newRule.urlPattern).toBe(urlPattern);
-    expect(newRule.isRegex).toBe(isRegex);
   });
 });
